Add ability to update dev tools from startup arguments
