+++ devtools/android/src/main/kotlin/com/maximbircu/devtools/android/DevToolsExtensions.kt
@@ -1,0 +1,12 @@
+package com.maximbircu.devtools.android
+
+import android.os.Bundle
+import com.maximbircu.devtools.common.DevTools
+
+fun DevTools.updateFromBundle(bundle: Bundle?) {
+    bundle?.getArguments()?.let(::updateFromParams)
+}
+
+private fun Bundle.getArguments(): Map<String, Any> = keySet().map { key ->
+    key to requireNotNull(get(key)) { "A value was not provided for $key" }
+}.toMap()
