+++ samples/android/src/main/kotlin/com/maximbircu/devtools/configscreens/YamlSourceConfigActivity.kt
@@ -7,15 +7,12 @@ import android.view.Gravity
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import com.maximbircu.devtools.R.layout
+import com.maximbircu.devtools.SampleApplication
 import com.maximbircu.devtools.android.DevToolsConfigurationScreen
-import com.maximbircu.devtools.android.readers.yaml
-import com.maximbircu.devtools.common.DevTools
-import com.maximbircu.devtools.common.readers.DevToolsSources
 import kotlinx.android.synthetic.main.activity_tools_container.devToolsContainer
 
-@Suppress("LongMethod")
 class YamlSourceConfigActivity : AppCompatActivity() {
-    private lateinit var devtools: DevTools
+    private val devtools = SampleApplication.application.yamlDevTools
 
     companion object {
         fun start(context: Context) {
@@ -27,9 +24,6 @@ class YamlSourceConfigActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(layout.activity_tools_container)
-
-        val source = DevToolsSources.yaml(assets, "dev-tools.yml")
-        devtools = DevTools.create("YAML", source)
 
         devtools.onConfigUpdated = { isCriticalUpdate ->
             val toast = Toast.makeText(
